pyproject.toml
[project]
name = "chess-arena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autopep8>=2.3.2",
    "fastapi>=0.115.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "python-chess>=1.999",
    "rich>=14.2.0",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
[tool.pycodestyle]
max-line-length = 120
ignore = ["E251", "E221"]

[tool.isort]
include_trailing_comma = false
use_parentheses = false
line_length = 120

[tool.coverage.run]
omit = ["*/test/*"]

